lambda/requirements.txt
ask-sdk-core>=1.18.0
google-generativeai>=0.5.0
boto3>=1.20.0
